requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dash==2.9.3
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.3.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.25.0
packaging==23.1
pandas==2.0.2
plotly==5.15.0
psycopg2-binary==2.9.6
py4j==0.10.9.7
pyspark==3.4.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
tenacity==8.2.2
tzdata==2023.3
urllib3==2.0.3
Werkzeug==2.3.6
